+++ src/managers/strategy/JecCommandStrategy.ts
@@ -1,0 +1,73 @@
+//  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+//
+//   Copyright 2016-2017 Pascal ECHEMANN.
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+/// <reference path="../../../typings/typings.d.ts"/>
+import {AbstractCommandStrategy} from "./AbstractCommandStrategy";
+import {CommandStrategy} from "./CommandStrategy";
+import * as chalk from "chalk";
+import * as figlet from "figlet";
+
+// Config file:
+const CFG:any = require("../../../config/jec-config.json");
+
+/**
+ * The <code>JecCommandStrategy</code> class invokes specific JEC commands
+ * depending on users inputs.
+ */
+export class JecCommandStrategy extends AbstractCommandStrategy
+                                        implements CommandStrategy {
+  
+  //////////////////////////////////////////////////////////////////////////////
+  // Constructor function
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * Creates a new <code>JecCommandStrategy</code> instance.
+   * 
+   * @param {string} version the version of the current process.
+   */
+  constructor(version:string) {
+    super(version);
+    this.initStrategy();
+  }
+  
+  //////////////////////////////////////////////////////////////////////////////
+  // Private properties
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * Initialises this object. This method is called by the constructor function.
+   */
+  private initStrategy():void {
+    console.log(
+      chalk.yellow(
+        figlet.textSync("JEC CLI", { horizontalLayout: "full" })
+      )
+    );
+    this.initCommands(CFG);
+  }
+  
+  //////////////////////////////////////////////////////////////////////////////
+  // Public methods
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * @inheritDoc
+   */
+  public invokeCommand():void {
+    super.invokeCommand();
+  }
+}