+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-# JEC CLI Project: Update Release Notess
+# JEC CLI Project: Update Release Notes
+
+<a name="jec-cli-0.0.2"></a>
+## **0.0.2** (2017-09-03)
+
+### Bug Fixes
+
+### Features
+
+- This release implements the JEC projects structure
+- Replacing `commander` by `minimist` as command parser
+- Adding the architechture for building dynamic commands based on the strategy pattern
 
 <a name="jec-cli-0.0.1"></a>
 ## **0.0.1** (2017-08-16)